testable(non-private) properties to be tested
In Swift, `private` means only within the same source file can access.
You can not even access it from the test. It has to be at least
internal, for it to be tested.

What's the use case for `private`? What's the major difference between
`internal` and `private`.

From the Swift Programming Language (Pre-release 2) book:
`internal` - App or Framework's internal structure.
`private` - to hide implementation.

Also see the note on page 875 in the book.